✨ feat(Header.tsx): Add Header component for page layout
A new file, Header.tsx, has been added to the client/src/pages directory. This file contains a functional component called Header, which is responsible for rendering the page header with a title and navigation links.

Changes made:
- Imported React module for JSX rendering.
- Defined the Props interface to specify the type of the 'title' prop.
- Created the Header functional component that takes the 'title' prop and renders the header HTML structure.
- Added a title element using the 'h1' tag with the value of the 'title' prop.
- Added a navigation section using the 'div' tag with the class name "navbar".
- Inside the navigation section, added two anchor tags for Home and ADD Note, respectively.

This new component can now be used to provide consistent header layout across different pages of the application.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>